% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checED50Range.R
\name{checkED50Range}
\alias{checkED50Range}
\title{in cases where estimated ED50 is out of range, checks whether the model fit should be shown or not}
\usage{
checkED50Range(
  fittedModel,
  inputData,
  slope,
  estimatedslope,
  inactiveSuperpotentParams
)
}
\arguments{
\item{fittedModel}{a drc object}

\item{inputData}{a data frame with two columns: column one the numeric dose, and column2 the response}

\item{slope}{the direction of the curve: either ascending or descending}

\item{estimatedslope}{the slope from the fitted model}

\item{inactiveSuperpotentParams}{a numeric vector of length 2. In cases where the response rage is below validRespRange, 
the first element of this vecotr shows the  response threshold to distinguish between inactive and suerpotent compounds, 
the second one shows the minimum proportion of response values that should be larger than this threshold to consider a compound superpotent. 
The default value for this argument is (50, 0.5) which means if more than half of data are larger than 50, then we consider the compound superpotent, 
and otherwise inactive.}
}
\value{
TRUE/FALSE
}
\description{
function to check checks whether the model fit should be shown or not
}
\author{
Vahid Nassiri
}
