% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compoundStatus.R
\name{compoundStatus}
\alias{compoundStatus}
\title{determines whether a compound is inactive or super potent}
\usage{
compoundStatus(
  inputData,
  inactiveSuperpotentParams,
  slope = c("ascending", "descending")
)
}
\arguments{
\item{inputData}{a data frame with two columns: column one the numeric dose, and column2 the response}

\item{inactiveSuperpotentParams}{a numeric vector of length 2. In cases where the response rage is below validRespRange, 
the first element of this vecotr shows the  response threshold to distinguish between inactive and suerpotent compounds, 
the second one shows the minimum proportion of response values that should be larger than this threshold to consider a compound superpotent. 
The default value for this argument is (50, 0.5) which means if more than half of data are larger than 50, then we consider the compound superpotent, 
and otherwise inactive.}

\item{slope}{the direction of the curve: either ascending or descending}
}
\value{
status of the compound
}
\description{
function to determine whether a compound if inactive or super potent
}
\author{
Vahid Nassiri
}
