% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determineTheSlope.R
\name{determineTheSlope}
\alias{determineTheSlope}
\title{determines the direction of the curve based on various factors}
\usage{
determineTheSlope(inputData, slope = NULL, estimatedSlope = NULL)
}
\arguments{
\item{inputData}{a data frame with two columns: column one the numeric dose, and column2 the response.}

\item{slope}{takes values from "ascending","descending", and "free", specifying the expected pattern.}

\item{estimatedSlope}{in case a model could be fitted, the estimated slope parameter.}
}
\value{

}
\description{
function to determine the direction of the curve
}
\author{
Vahid Nassiri
}
