% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateRange.R
\name{estimateRange}
\alias{estimateRange}
\title{computes the estimated range}
\usage{
estimateRange(fittedModel, fixedBottom, fixedTop)
}
\arguments{
\item{fittedModel}{a drc object}

\item{fixedBottom}{fixed bottom}

\item{fixedTop}{fixed top}
}
\description{
function to compute the estimated rage
}
\author{
Vahid Nassiri
}
