% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractLogisticModelFit.R
\name{extractLogisticModelFit}
\alias{extractLogisticModelFit}
\title{extracts required stuff from a fitted drc model}
\usage{
extractLogisticModelFit(
  fittedModel,
  inputData,
  coef2return,
  responseName,
  confLevel,
  slope = c("ascending", "descending"),
  inactiveSuperpotentParams,
  inputDataPlot,
  respLevpICx,
  nPointsPredict
)
}
\arguments{
\item{fittedModel}{a drc object}

\item{inputData}{a data frame with two columns: column one the numeric dose, and column2 the response}

\item{coef2return}{scheme of the estimated coeffiient to return.}

\item{responseName}{string specifying the response name to be diplayed on the plot}

\item{confLevel}{confidence level for the confidence intervals and significance tests, default is 0.95.}

\item{slope}{the slope of the curve: either "ascending" or "descending"}

\item{inactiveSuperpotentParams}{a numeric vector of length 2. In cases where the response rage is below validRespRange, 
the first element of this vecotr shows the  response threshold to distinguish between inactive and suerpotent compounds, 
the second one shows the minimum proportion of response values that should be larger than this threshold to consider a compound superpotent. 
The default value for this argument is (50, 0.5) which means if more than half of data are larger than 50, then we consider the compound superpotent, 
and otherwise inactive.}

\item{inputDataPlot}{the complete inputData with accept column}

\item{respLevpICx}{scalar between 0 and 100 to specify response level for computing user-speficied pICx, for example  respLevpICx = 30 will return pIC30.}

\item{nPointsPredict}{scalar specifying the number of points to be used to make fitted curve with its confidence bands}
}
\description{
function to extract required stuff from a fitted drc model
}
\author{
Vahid Nassiri
}
