% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fittingLogisticModel.R
\name{fittingLogisticModel}
\alias{fittingLogisticModel}
\title{fits a logistic model to the dose-response data}
\usage{
fittingLogisticModel(
  inputData,
  fixedBottom = NA,
  fixedTop = NA,
  fixedSlope = NA,
  confLevel = 0.95,
  robustMethod = c("mean", "tukey", "median"),
  responseName,
  slope = c("ascending", "descending", "free"),
  validRespRange = 30,
  receptorStyleModel = c("PL4", "PL3L", "PL3U", "2PL", "PL4H1", "PL3LH1", "PL3UH1",
    "PL2H1"),
  inactiveSuperpotentParams = c(50, 0.5),
  accept = NULL,
  respLevpICx = NULL,
  nPointsPredict = 100
)
}
\arguments{
\item{inputData}{a data frame with two columns: column one the numeric dose, and column2 the response}

\item{fixedBottom}{the fixed lower limit parameter of the logistic model, default is NULL, if provided then the lower limit will be fixed.}

\item{fixedTop}{the fixed upper limit parameter of the logistic model, default is NULL, if provided then the upper limit will be fixed.}

\item{fixedSlope}{the fixed hill slope parameter of the logistic model, default is NULL, if provided then the hill slope will be fixed.}

\item{confLevel}{confidence level for the confidence intervals and significance tests, default is 0.95.}

\item{robustMethod}{character with three possibilities: mean for the non-robust fit (default) and tukey and median for the robust fit.}

\item{responseName}{string specifying the response name to be diplayed on the plot}

\item{slope}{takes values from "ascending","descending", "free", specifying the expected pattern.}

\item{validRespRange}{valid difference of minimum and maximum response (default is 30), if the range is smaller than this, no analysis will be done.}

\item{receptorStyleModel}{the general model to fit is a logistic model, and one can fix different parameters of it via fixedBottom, 
fixedTop, and fixedSlope. However, if case one is intersted to use the style in receptor package, then the model can be specified here. One may selected a model from  "PL4", "PL3L", "PL3U", "2PL", "PL4H1", "PL3LH1", "PL3UH1", "PL2H1", similar to what we have in receptor. 
If this argument is left empty, then the drc concentions are used.}

\item{inactiveSuperpotentParams}{a numeric vector of length 2. In cases where the response rage is below validRespRange, 
the first element of this vecotr shows the  response threshold to distinguish between inactive and suerpotent compounds, 
the second one shows the minimum proportion of response values that should be larger than this threshold to consider a compound superpotent. 
The default value for this argument is (50, 0.5) which means if more than half of data are larger than 50, then we consider the compound superpotent, 
and otherwise inactive.}

\item{accept}{a vector the same length as the sample size (including replications) with 1's and 0's. Zero's will correspond to the points that will excluded 
from the analysis. If nothing is given then all the points are used.}

\item{respLevpICx}{scalar between 0 and 100 to specify response level for computing user-speficied pICx, for example  respLevpICx = 30 will return pIC30.}

\item{nPointsPredict}{scalar specifying the number of points to be used to make fitted curve with its confidence bands}
}
\value{
a list with the following objects
respIC50 a vector with raw estimated pIC50 results
validpIC50 a vector with valid pIC50 results (after sanity checks)
checkpIC50 results of sanity checks
respIC20 a vector with raw estimated pIC20 results
validpIC20 a vector with valid pIC20 results (after sanity checks)
checkpIC20 results of sanity checks
respIC80 a vector with raw estimated pIC80 results
validpIC80 a vector with valid pIC80 results (after sanity checks)
checkpIC80 results of sanity checks
respICx a vector with raw estimated pICx results
validpICx a vector with valid pICx results (after sanity checks)
checkpICx results of sanity checks
modelCoefs a matrix with model parameter estimates, standard errors, significance tests and confidence intervals.
residulaVariance scalar providing error variance
warningPD character showing if the parameter covariance matrix is positive definite or not.
fittedModel a drc object of the fitted model
plot a ggplot2 object with the plotted data, fitted model, confidence bands and pIC50.
AIC, AICc and BIC of the fitted model.
dataPredict2Plot a data frame with predicted values for the curve and its confidence bands (if they can estimated)
pIC50Location a vector of length 2 with the coordinates of pIC50 on x and y axis for plotting (if they can be estimated)
xAxisLabels as the doses are transformed a couple of times, the real values on x-axis are different from the ones which are plotted, so to make it easier for plotting later on the labels we used are also provided.
xAxisBreaks x axix breaks used to produce the plot are also provided.
weights if robustMethod = "tukey" then the obtained weights are returned. The same weights are used to make the plots (size of the points).
tukeyMessage a message showing the convergence situation of the tukey biweight (obustMethod = "tukey")
}
\description{
function to fit a logisitic model to the data already in log-sclae
}
\author{
Vahid Nassiri
}
