% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isSaneEDx.R
\name{isSaneEDx}
\alias{isSaneEDx}
\title{checks sanity conditions for an estimated ED50}
\usage{
isSaneEDx(resEDx, doseVals)
}
\arguments{
\item{resEDx}{a vector with estimate, standard error and confidence interval of the estimated effective dose}

\item{doseVals}{the dose values}
}
\value{
a vector of length 5 with the results of 5 sanity check.
}
\description{
function to check sanity of an estimated EDx
}
\author{
Vahid Nassiri
}
