% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makingDoseResponseData.R
\name{makingDoseResponseData}
\alias{makingDoseResponseData}
\title{creates suitable dose-response data out of the provided Excel files}
\usage{
makingDoseResponseData(assayData, aggregatePlates = TRUE)
}
\arguments{
\item{assayData}{the outcome of the Excel file.}
}
\value{
a list with suitabel dose-response data frames as its elements.
}
\description{
function to create suitabel data frames with dose-response data to be used in fittingLogisticModel function.
}
\author{
Vahid Nassiri
}
