% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strEndsWith.R
\name{strEndsWith}
\alias{strEndsWith}
\title{checks if a strings ends with certain characters}
\usage{
strEndsWith(haystack, needle)
}
\arguments{
\item{haystack}{the main string}

\item{needle}{the ending string}
}
\value{
TRUE or FALSE
}
\description{
checks if a strings ends with certain characters
}
\author{
https://stackoverflow.com/questions/26188663/how-to-determine-if-a-string-ends-with-another-string-in-r
}
