% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformEstimates.R
\name{transformEstimates}
\alias{transformEstimates}
\title{transforms the estimated ICx's in natrual log to log10.}
\usage{
transformEstimates(estVec, fittedModel)
}
\arguments{
\item{estVec}{vector of length 4 with estimate, standard error and confidence interval.}

\item{fittedModel}{drc opbject of the fitted model to obtain degrees of freedom of the residual.}
}
\value{
a vector of length 4 with estimate, standard error and confidence interval for the transformed ICx.
}
\description{
transforms the estimated ICx's in natrual log to log10.
}
\author{
Vahid Nassiri
}
