% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validEDx.R
\name{validEDx}
\alias{validEDx}
\title{returns the valid estimated EDx results}
\usage{
validEDx(resEDx, doseVals, p)
}
\arguments{
\item{resEDx}{a vector with estimate, standard error and confidence interval of the estimated effective dose}

\item{doseVals}{the dose values}

\item{p}{effective dose level (between 0 and 100, e.g. p = 50 for IC50)}
}
\value{
a vector with valid elements.
}
\description{
function to return the valid estimated EDx results
}
\author{
Vahid Nassiri
}
